+++ asio-sys/src/bindings/mod.rs
@@ -924,6 +924,11 @@ impl Driver {
         let mut dcb = DRIVER_EVENT_CALLBACKS.lock().unwrap();
         dcb.retain(|&(id, _)| id != rem_id);
     }
+
+    /// Opens the ASIO driver's control panel window.
+    pub fn open_control_panel(&self) -> Result<(), AsioError> {
+        unsafe { asio_result!(ai::ASIOControlPanel()) }
+    }
 }
 
 impl DriverState {